`HpCalculation`: fix error message for exit code 495 (#93)
Fixes #92

The exit code 495 message speaks about problems with the "parallelization options".
This is not the case, as the problem is only due an incompatible FFT grid with the
symmetries found. This can happen at the perturbative level when a structure is close to
an other space group symmetry, leading to noise in their detection and usage within the
linear-response package of QE. This can in principle be fixed by using the corresponding
ibrav, by swithing off symmetries (not recommended), or by relaxing better the structure.
Automated handling of such case can be implemented in the future at higher level in the
SelfConsistentHubbardWorkChain.