🐛 Fix incorrect assignement of pseudos Fixes #107
`re.match()` only requires the pattern to match from the start;
it does not require consuming the whole string, making
element couples such as (`H`, `Hg`)  and (`C`,`Co`) to
be assigned to the same pseudo. Now we match the full
name and add a test for these limiting cases.