👌 Parser: add exit codes for diverging Hubbard parameters (#110)
Fixes #102

When the value of an Hubbard parameter is too large,
fortran converts it into a string such as `*****`. This
makes it impossible to parse. At the same time, it is
a signature that the Hubbard parameter is diverging,
and therefore one should stop the self-consistent
cycle as early as possible.